wip to split and refactor the validators into three separate vlaidators
